reorganised test_views & created custom TestCase
